% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.gwas-package.R
\docType{package}
\name{sp.gwas-package}
\alias{sp.gwas-package}
\title{Selection probabilities using generalized linear model with regularization for a SNP data in the hapmap format.}
\value{
A list of data files(genotype, phenotype, etc.), results for selection probabilities, and manhattan plot for multiple traits.
}
\description{
sp.gwas
}
\details{
The penalty function of \code{elastic-net} is defined as \deqn{\alpha||\beta||_1+(1-\alpha)||\beta||_2/2,} where \eqn{\alpha} is a mixing proportion of ridge and the lasso, and \eqn{\beta} is regression coefficients. This penalty is equivalent to the Lasso penalty if \code{alpha=1}. \cr
}
\references{
Zou, H., & Hastie, T. (2005). Regularization and variable selection via the elastic net. Journal of the royal statistical society: series B (statistical methodology), 67(2), 301-320.
}
\keyword{"selection}
\keyword{gwas}
\keyword{hapmap}
\keyword{probabilities"}
\keyword{regularization}
