% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.gwas.R
\name{sp.gwas}
\alias{sp.gwas}
\title{Selection probabilities using generalized linear model with regularization for a SNP data in the hapmap format.}
\usage{
sp.gwas(genotype.path = NULL, phenotype.path = NULL,
  save.path = "./sp.folder", y.col = 4, y.id.col = 2,
  method = "lasso", family = "gaussian", Falsediscovery = c(1, 5,
  10), plot.ylim = NULL, lambda.min.quantile = 0.5, n.lambda = 10,
  K = 100, psub = 0.5, plot.name = "", plot.type = "jpg",
  plot.dpi = 300)
}
\arguments{
\item{genotype.path}{A path of a snp data which is a \code{p} by \code{(n+11)} matrix of genotypes. It is formatted by hapmap which has (rs, allele, chr, pos) in the first four(1-4) columns, (strand, assembly, center, protLSID, assayLSID, panel, Qcode) in the following seven(5-11) columns. If NULL, user can choose a path in interactive use.}

\item{phenotype.path}{A path of a phenotype data which is an \code{n} by \code{p} matrix of phenotypes. Since the first some columns can display attributes of the phenotypes, you should enter the arguments, y.col and y.id.col, which represent the columns of phenotypes to be analyzed and the column of sample ID. If NULL, user can choose a path in interactive use.}

\item{save.path}{A save.path which has all output files. If there exists save.path, sp.gwas will check if there is an output file. If then, sp.gwas is not going to generate the results but just loading the output files(.RData).}

\item{y.col}{The columns of phenotypes. At most 4 phenotypes can be considered, because the plot of them will be fine. Default is 4.}

\item{y.id.col}{The column of sample ID in the phenotype data file. Default is 2.}

\item{method}{A method of penalized regression. It includes "lasso" for the lasso and "enet" for the elastic-net.}

\item{family}{A family of response variable(phenotype). It is "gaussian" for continuous response variable, "binomial" for binary, "poisson" for count, etc. For more details, see the function(\code{stats::glm}). Default is "gaussian".}

\item{Falsediscovery}{The expected number of false discovery to be controlled. The larger it is, the higher threshold becomes. Default is c(1, 5, 10).}

\item{plot.ylim}{A range of the y-axis. If NULL, automatic range in the y-axis will be provided. For plot.ylim=c(0,1), the y-axis has a range of 0 and 1.}

\item{lambda.min.quantile}{A range of lambda sequence. Default is 0.5 (median). If the range is so small that it can have many tied selection probabilities which is 1.}

\item{n.lambda}{The length of lambda sequence. The larger n.lambda, the more detailed lambda sequence will be.}

\item{K}{The number of iterations in resampling when calculating the selection probabilities.}

\item{psub}{The subsampling proportion. For efficiency, default is 0.5.}

\item{plot.name}{A name of plot file.}

\item{plot.type}{A type of plot file which includes "jpg", "pdf", "tiff", etc.}

\item{plot.dpi}{A resolution of plot. If you want to get a high-resolution image, plot.dpi should be large.}
}
\value{
\item{Histogram of original and transformed phenotypes}{Histogram of phenotypes with p-value by Shapiro-test on the top right corner.}
    \item{myDATA}{A list of myX, myGD, myGM, myGT, myY, and myY.t.}
    \item{sp.res}{A list of sp.df and threshold.}
    \item{Circular Manhattan plot}{Manhattan plot for the first phenotype is the innermost circle. Colors for chromosome is fixed, so that if you want to change colors, you would edit the R code of sp.manhattan function.}
}
\description{
For analysis of high-dimensional genomic data, penalized regression can be a solution to accomodate correlations between predictors. Moreover, selection probabilities do not depend on tuning parameter selection so that it produces a stablity selection. Thresholds are also generated to control the false positives(errors). Thresholds vary with the expected number of false positives to be controlled by the user.
Input data is the hapmap formatted SNP data and phenotype data corresponding to SNP data.
Output includes files of three types: (1) Matched data files (genotype, numerical, snp info, and phenotype), (2) Results file (selection probabilities and thresholds), (3) Circular manhattan plot with (blue dotted) significant line corresponding to the largest value among user-defined false discoveries.
}
\details{
The penalty function of \code{pclogit} is defined as \deqn{\alpha||\beta||_1+(1-\alpha)||\beta||_2/2,} where \eqn{\alpha} is a mixing proportion of ridge and the lasso, and \eqn{\beta} is regression coefficients. This penalty is equivalent to the Lasso penalty if \code{alpha=1}. \cr
}
\examples{
# Not run
# sp.gwas(genotype.path = "./input.snp.csv",
#         phenotype.path = "./input.phenotype.csv",
#         save.path = "./Test",
#         y.col=5:8,
#         y.id.col=2,
#         method="lasso",
#         family="gaussian",
#         Falsediscovery = c(1,5,10),
#         plot.ylim = NULL,
#         lambda.min.quantile = 0.5,
#         n.lambda = 10,
#         K = 100,
#         psub = 0.5,
#         plot.name = "Test",
#         plot.type = "jpg",
#         plot.dpi = 300)
}
\author{
Kipoong Kim <kkp7700@gmail.com>
}
